Create UI.js
